+++ src/components/Table/index.ts
@@ -1,3 +1,3 @@
-// Table from copy-ui @ 2025-11-23.
+// Table from copy-ui @ 2025-11-25.
 
 export { Table } from "./table";